services/matching-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
redis>=5.0.0
confluent-kafka>=2.4.0
python-geohash>=0.8.5
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
python-multipart>=0.0.9
email-validator>=2.1.0
